requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
Django==2.2.5
django-cors-headers==3.1.0
django-filter==2.2.0
djangorestframework==3.10.3
entrypoints==0.3
factory-boy==2.12.0
Faker==2.0.1
flake8==3.7.8
importlib-metadata==0.21
inflection==0.3.1
invoke==1.3.0
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.1
pluggy==0.12.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.2
pytest==5.1.2
pytest-django==3.5.1
pytest-factoryboy==2.0.3
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
wcwidth==0.1.7
zipp==0.6.0
